refactor: move checklist widgets to subfolder
