+++ lib/features/tasks/ui/checklists/checklist_widget.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:lotti/features/tasks/ui/checklist_item_wrapper.dart';
-import 'package:lotti/features/tasks/ui/consts.dart';
+import 'package:lotti/features/tasks/ui/checklists/checklist_item_wrapper.dart';
+import 'package:lotti/features/tasks/ui/checklists/consts.dart';
 import 'package:lotti/features/tasks/ui/title_text_field.dart';
 import 'package:lotti/l10n/app_localizations_context.dart';
 import 'package:lotti/themes/colors.dart';
